requirements.txt
groq
SpeechRecognition
pyttsx3
pyaudio
python-dotenv
streamlit

